Update index.md
I propose removing the word "long" from the BOM in this section.  On the eccentric spacer packets, the longer spacers have the word "long" on the label, but the 0.200" spacers do not, so it appears that the word "long" is being used to identify the longer set of spacers rather than specifying that they are 0.375" long.